==> 1_Crowdflower_data.ipynb <==
import pandas as pd
import numpy as np
import re

pd.set_option('display.max_colwidth', -1)

cf_twitter_data = pd.read_csv("data/twitter_gender.csv", encoding='latin1')

cf_twitter_data.sample(2)

cf_twitter_data.columns

len(cf_twitter_data['_unit_id'])

# how many where confidence in gender prediction is above 70%?
len(cf_twitter_data[cf_twitter_data['gender:confidence'] >= 0.7])

# select only rows where gender confidence is above 70%, and gender is male or female (NOT brand)
gender_conf_df = cf_twitter_data[
    (cf_twitter_data['gender:confidence'] >= 0.7) & 
    (
        (cf_twitter_data['gender'] == 'male') |
        (cf_twitter_data['gender'] == 'female')
    )
]

gender_conf_df['name'][:10]

len(gender_conf_df['_unit_id'])

# make list of usernames
usernames = list(gender_conf_df['name'])

len(usernames)

type(usernames)

np.save('usernames_list', usernames)

# save df of userids and genders

userid_gender_df = gender_conf_df[['_unit_id', 'gender']]

userid_gender_df.head()

pd.to_pickle(userid_gender_df, 'userid_gender_df')
