==> 3_Clean_tweets.ipynb <==
import pymongo
import pickle
import pandas as pd
pd.set_option('display.max_colwidth', -1)
import numpy as np
import re
import csv

import matplotlib.pyplot as plt
%matplotlib inline

import nltk
from sklearn.feature_extraction.text import CountVectorizer
from wordcloud import WordCloud, STOPWORDS

def cleaner(string):
    string = re.sub('http*\S*', '', string)
    string = re.sub('amp;', '', string)
    string = re.sub('\n', ' ', string)
    return string

alltweets_df = pd.read_csv('tweets.csv', names=['user_name', 'tw_id', 'text'], encoding='utf-8')

alltweets_df['tw_id'].count()

# drop duplicates on tw_id and text
alltweets_df = alltweets_df.drop_duplicates('tw_id')

alltweets_df['tw_id'].count()

alltweets_df = alltweets_df.drop_duplicates('text')

alltweets_df['tw_id'].count()

# drop retweets
alltweets_df = alltweets_df[~alltweets_df['text'].str.startswith('RT')]

alltweets_df['tw_id'].count()

alltweets_df.sample(40)

# clean text
alltweets_df['text'] = alltweets_df['text'].apply(cleaner)

alltweets_df.sample(50)

# Drop non-English tweets - HOW?

def drop_non_english(string):
    alpha_emojis = ['a', 'b', 'c', 'd', ]
    for letter in string:
        

alltweets_df.sort('text', ascending=False)[:150]

gender_df = pd.read_pickle('crowd_flower_gender_df')

# add gender column to tweet_df, joining on 'user_name'
alltweets_df = pd.merge(alltweets_df, gender_df, how='left', on='user_name')

alltweets_df.sample(10)

alltweets_df['gender'].value_counts()

# drop brand and unknown
m_f_tweets_df = alltweets_df[(alltweets_df['gender']=='male') | (alltweets_df['gender']=='female')]

m_f_tweets_df['gender'].value_counts()

# pickle the df
m_f_tweets_df.to_pickle('m_f_tweets_df')
