==> 4_Emojis_to_text.ipynb <==
import pickle
import pandas as pd
pd.set_option('display.max_colwidth', -1)
import numpy as np
import re
import csv
import matplotlib.pyplot as plt
%matplotlib inline

# load m_f_tweets_df

m_f_tweets_df = pd.read_pickle('m_f_tweets_df')

# don't know where duplicates came from; drop
m_f_tweets_df = m_f_tweets_df.drop_duplicates('tw_id')

m_f_tweets_df['gender'].count()

tweet_text = np.asarray(m_f_tweets_df['text'])
gender = np.asarray(m_f_tweets_df['gender'])

# pickle gender
np.save('gender.npy', gender)

len(tweet_text)

len(gender)

# check that the classes are balanced

np.unique(gender, return_counts=True)

Get a list of all emojis that exist:

with open('emojis.txt', 'r') as f:
    emojis = f.read()

emojis = emojis.split(sep='\n')

len(emojis)

# make a list of all emojis that appear in the tweets

def make_emoji_list(tweets, emojis):
    emojis_in_tweets = []

    for emoji in emojis:
        for sentence in tweets:
            if emoji in sentence:
                emojis_in_tweets.append(emoji)
    
    return emojis_in_tweets           

emojis_in_tweets = make_emoji_list(tweet_text, emojis)

len(emojis_in_tweets)

emoji_string = ' '.join(emojis_in_tweets)

len(emoji_string)

with open("emoji_string.txt", "w") as f:
    f.write(emoji_string)

# make dict of emojis and counts
from collections import Counter
emoji_count = Counter(emojis_in_tweets)

sorted_emojis = emoji_count.most_common()

top_50 = sorted_emojis[:50]

counts = []
emojis = []

for i in top_50:
    emojis.append(i[0])
    counts.append(i[1])

"""# Set the bar labels
bar_labels = emojis[:25]

#define plot size in inches (width, height) & resolution(DPI)
fig = plt.figure(figsize=(10, 6), dpi=100)

# Create the x position of the bars
x_pos = list(range(len(bar_labels)))

# Create the plot bars
# In x position
plt.bar(x_pos,
        # using the data from the counts
        counts[:25],
        # aligned in the center
        align='center',
        # with color
        color='indigo',
        # alpha 0.5
        alpha=0.5)


# set height of the y-axis
#max_y = max(zip(mean_values, variance)) # returns a tuple, here: (3, 5)
#plt.ylim([0, (max_y[0] + max_y[1]) * 1.1])

# set axes labels and title
plt.ylabel('Count')
plt.xticks(x_pos, bar_labels)
plt.title('Top 25 emojis in all tweets')

plt.show()"""

Make a dictionary of emojis and (synthetic) emoji names:

def emoji_to_word(emojis):
    emoji_dict = {}
    for i, emoji in enumerate(emojis):
        emoji_dict[emoji] = 'emoji_' + str(i)
    return emoji_dict

emoji_dict = emoji_to_word(emojis)

# pickle emoji_dict

with open('emoji_dict.pickle', 'wb') as handle:
    pickle.dump(emoji_dict, handle, protocol=pickle.HIGHEST_PROTOCOL)

Use emoji_dict to replace emojis with their words in tweet text (for vectorizing):

def replace_emojis(tweets):
    new_tweets = []
    for sentence in tweets:
        for emoji in emojis:
            if emoji in sentence:
                sentence = sentence.replace(emoji, emoji_dict[emoji] + ' ')
        new_tweets.append(sentence)
    return new_tweets

tweets_emoji_words = replace_emojis(tweet_text)

len(tweets_emoji_words)

# save tweets_emoji_words

with open('tweets_emoji_words.pkl', 'wb') as f:
    pickle.dump(tweets_emoji_words, f)
