==> 5_NLP.ipynb <==
import pymongo
import pickle
import pandas as pd
pd.set_option('display.max_colwidth', -1)
import numpy as np
import csv

import matplotlib.pyplot as plt
%matplotlib inline
from wordcloud import WordCloud, STOPWORDS
import seaborn as sns
from nltk.corpus import stopwords

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.feature_extraction import text
from sklearn.decomposition import TruncatedSVD
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV, StratifiedShuffleSplit
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression

with open('tweets_emoji_words.pkl', 'rb') as f:
    tweets_emoji_words = pickle.load(f)

len(tweets_emoji_words)

tweet_text = tweets_emoji_words

# remove non-English from text
#tweet_text = [char.encode('ascii', errors='ignore') for char in tweet_text]

len(tweet_text)

# get gender
gender = np.load('gender.npy')

len(gender)

tweets_gender_df = pd.DataFrame(
    {'text': tweet_text,
     'gender': gender
    })

fem_text = tweets_gender_df[tweets_gender_df['gender'] == 'female']['text']

male_text = tweets_gender_df[tweets_gender_df['gender'] == 'male']['text']

X_train, X_test, y_train, y_test = train_test_split(
    tweet_text, gender, test_size=0.33, random_state=42)

y = gender

# split the training data into 'M' and 'F' 
X_tr_M = X_train[y_train=='male']

X_tr_F = X_train[y_train=='female']

What are the top 25 'words' for female and for male?

from collections import Counter

stop_words = text.ENGLISH_STOP_WORDS

# in: pandas series of strings

def series_to_word_counts(series):

    word_counts = Counter(' '.join(series.astype('str').values.tolist()).split())
    word_counts = list(word_counts.items())
    word_counts = [(i.lower(), j) for i, j in word_counts]
    word_counts.sort(key=lambda x: x[1], reverse=True)
    
    return word_counts
# out: sorted list of word counts (lower case)

# in: sorted list of word counts

def top_25_word_counts(word_counts):
    total_counter = 0
    non_stop_counter = 0
    word_counts_list = []

    while non_stop_counter < 25:
        if word_counts[total_counter][0] not in stop_words:
            word_counts_list.append(word_counts[total_counter])
            non_stop_counter += 1
        total_counter += 1
        
    return word_counts_list
# out: top 25 words + counts

fem_word_counts = series_to_word_counts(fem_text)

fem_word_list = top_25_word_counts(fem_word_counts)

top_25_fem = [i for i,j in fem_word_list]

top_25_fem

male_word_counts = series_to_word_counts(male_text)

male_word_list = top_25_word_counts(male_word_counts)

top_25_male = [i for i,j in male_word_list]

top_25_male

# get original emojis from emoji dict and replace in list

# open emoji_dict 
with open('emoji_dict.pickle', 'rb') as handle:
    emoji_dict = pickle.load(handle)

# reverse the dict
emoji_word_dict = { v:k for k,v in emoji_dict.items() }

def replace_emoji_words(words):
    for word in words:
        word = emoji_word_dict[word]

top_25_fem = [emoji_word_dict.get(item,item)  for item in top_25_fem]

top_25_fem

top_25_male = [emoji_word_dict.get(item,item)  for item in top_25_male]

top_25_male

Count Vectorize:

cv = CountVectorizer(stop_words='english', min_df=10)

cv.fit(X_train)

X_cv = cv.transform(X_train)

X_cv.shape

Make a pipeline: Tfidf --> SVD --> Log Reg, Random Forest

from sklearn.pipeline import Pipeline

vec_pipe = Pipeline([('tfidf', TfidfVectorizer(min_df = 3, stop_words = 'english')),
                     ('svd', TruncatedSVD(n_components=200)),
                     ('lr', LogisticRegression()),
                    ])

text_clf = text_clf.fit(twenty_train.data, twenty_train.target)

Tf-idf:

tfidf = TfidfVectorizer(min_df = 5, stop_words = 'english')

X_tr_tf = tfidf.fit_transform(X_train)

X_tr_tf

X_test_tf = tfidf.transform(X_test)

features_tf = tfidf.get_feature_names()

features_tf

X_tr_tf.shape

SVD = TruncatedSVD(n_components=2)
LSA_tr = SVD.fit_transform(X_tr_tf)

SVD.components_

LSA_test = SVD.transform(X_test_tf)

component_names = ["component_"+str(i+1) for i in range(2)]

LSA_df = pd.DataFrame(SVD.components_,
                                         index = component_names,
                                         columns = tfidf.get_feature_names()).T

fem_top_20_feats = LSA_df.sort('component_1', ascending=False)[:20]

m_top_20_feats = LSA_df.sort('component_2', ascending=False)[:20]

def replace_emoji_words(item):
    return emoji_word_dict.get(item,item)

fem_top_20_feats.index = fem_top_20_feats.index.map(replace_emoji_words)

fem_top_20_feats.iloc[:,:1]

m_top_20_feats.index = m_top_20_feats.index.map(replace_emoji_words)

m_top_20_feats.iloc[:,1:2]

document_term_matrix = pd.DataFrame(X_tr_tf.toarray(),
                                        index=X_train,
                                        columns=tfidf.get_feature_names ())

document_term_matrix.head()

Classification Models:

param_dict = {
    'lr':{
        'C':np.logspace(-3,3,7)
    },
    'rf':{}
}

model_dict = {
    'lr':GridSearchCV(LogisticRegression(),
                             param_grid=param_dict['lr'],
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
    'rf':GridSearchCV(RandomForestClassifier(),
                      param_grid=param_dict['rf'],
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42)),
}

def fit_all_models(x,y, model_dict):
    for model in model_dict.keys():
        model_dict[model].fit(x,y)
        print("{:5} best score: {}".format(model, model_dict[model].best_score_))

fit_all_models(LSA_tr, y_train, model_dict)

lr = GridSearchCV(LogisticRegression(),
                             param_grid=param_dict['lr'],
                             cv=StratifiedShuffleSplit(n_splits=5, random_state=42))

lr.fit(LSA_tr, y_train)

lr.cv_results_

rf = GridSearchCV(RandomForestClassifier(),
                      param_grid=param_dict['rf'],
                      cv=StratifiedShuffleSplit(n_splits=5, random_state=42))

rf.fit(LSA_tr, y_train)

rf.cv_results_

results = pd.DataFrame(rf.cv_results_)
results

Predict gender on test set:

predicted_gender_lr = model_dict[0].predict(LSA_test)

predicted_gender_rf = model_dict[1].predict(LSA_test)

dfs = top_feats_by_class(X_tf, y, features)

plot_tfidf_classfeats_h(dfs)

From Friday afternoon with Josh and Shehreen:

from sklearn.pipeline import Pipeline

tfd_pipe = Pipeline([
    ('tfidf', TfidfVectorizer()),
    ('svd', TruncatedSVD(n_components=6))
])

tfd_pipe.fit(cleantweets)

tfd = tfd_pipe.steps[0][1]
svd = tfd_pipe.steps[1][1]

tfd, svd

tfd_word_index = tfd.get_feature_names()

svd.components_.shape

import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
def svd_variance(svd_model, col_index):
    
    dimensions = ['Dimension {}'.format(i) for i in range(1,len(svd_model.components_)+1)]
    components = pd.DataFrame(np.round(svd_model.components_, 4), columns=col_index)
    ratios = svd_model.explained_variance_ratio_.reshape(len(svd_model.components_), 1)
    variance_ratios = pd.DataFrame(np.round(ratios, 4), columns = ['Explained Variance'])
    variance_ratios.index = dimensions

    fig, ax = plt.subplots(figsize = (24,10))

    # Plot the feature weights as a function of the components
    components.plot(ax = ax, kind = 'bar',legend='False');
    ax.set_ylabel("Feature Weights")
    ax.set_xticklabels(dimensions, rotation=0)
    ax.legend_.remove()


    # Display the explained variance ratios
    for i, ev in enumerate(svd_model.explained_variance_ratio_):
        ax.text(i-0.40, ax.get_ylim()[1] + 0.05, "Explained Variance\n          %.4f"%(ev))

svd_variance(svd, tfd_word_index)

svd

[word for word
in sorted(zip(tfd_word_index, svd.components_[0]), key = lambda x: x[1], reverse=True)
         if abs(word[1]) > .3]

[word for word
in sorted(zip(tfd_word_index, svd.components_[1]), key = lambda x: x[1], reverse=True)
         if abs(word[1]) > .3]
